Move to GetFullPathUsingSystemPathOrDefault to a standalone file SystemPathSearcher
