+++ MedallionShell/Shell.cs
@@ -6,7 +6,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -39,7 +38,7 @@ namespace Medallion.Shell
             Throw.If(string.IsNullOrEmpty(executable), "executable is required");
 
             var executablePath = !executable.Contains(Path.DirectorySeparatorChar)
-                && GetFullPathUsingSystemPathOrDefault(executable) is { } fullPath
+                && SystemPathSearcher.GetFullPathUsingSystemPathOrDefault(executable) is { } fullPath
                 ? fullPath
                 : executable;
 
@@ -81,24 +80,6 @@ namespace Medallion.Shell
             }
             
             return command;
-        }
-
-        // internal for testing
-        internal static string? GetFullPathUsingSystemPathOrDefault(string executable)
-        {
-            var paths = Environment.GetEnvironmentVariable("PATH")!.Split(Path.PathSeparator);
-
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                var pathExtensions = Environment.GetEnvironmentVariable("PATHEXT")!
-                    .Split(Path.PathSeparator)
-                    .Concat(new[] { string.Empty })
-                    .ToArray();
-                return paths.SelectMany(path => pathExtensions.Select(pathExtension => Path.Combine(path, executable + pathExtension)))
-                    .FirstOrDefault(File.Exists);
-            }
-
-            return paths.Select(path => Path.Combine(path, executable)).FirstOrDefault(File.Exists);
         }
 
         private static void PopulateArguments(ProcessStartInfo processStartInfo, IEnumerable<object?>? arguments, Options options)
